% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covertree.R
\name{covertree}
\alias{covertree}
\title{Cover Tree}
\usage{
covertree(data, scalefactor = 2, metric = dist2)
}
\arguments{
\item{data}{Matrix of data in rows}

\item{scalefactor}{Zooming factor for the cover tree}

\item{metric}{Function used to compute distances}
}
\description{
TODO: Implement insert, remove, and tree balancing
}

