% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmra_ct.R
\name{covertree_GMRA}
\alias{covertree_GMRA}
\title{GMRA from cover trees}
\usage{
covertree_GMRA(ct, sc, dim)
}
\arguments{
\item{ct}{The cover tree used to build the GMRA}

\item{sc}{The scale for determing the cover tree partition}

\item{dim}{The rank of the local projections}
}
\description{
TODO: Usage of eigen is expensive; fast option needed
}

