% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmra_ct.R
\name{projector}
\alias{projector}
\title{Orthogonal projection}
\usage{
projector(X, mu, dim)
}
\arguments{
\item{A}{matrix with data in rows}

\item{The}{shift value to be subtracted from each row of X}

\item{The}{rank of the projection to be extracted}
}
\description{
TODO: Implement randomized low-rank SVD approach
}

