% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{runifsphere}
\alias{runifsphere}
\title{Uniform random samples from unit sphere}
\usage{
runifsphere(n, dim = 2)
}
\arguments{
\item{n}{Number of samples}

\item{dim}{Dimension of the Euclidean space}
}
\description{
Uniform random samples from unit sphere
}

