% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmra_ct.R
\name{projector}
\alias{projector}
\title{Orthogonal projection}
\usage{
projector(A, dim)
}
\arguments{
\item{A}{Symmetric matrix}

\item{dim}{Number of eigenvectors to retain}
}
\description{
TODO: Usage of eigen is expensive; fast option needed
}

